LayoutTests/fast/dom/Document/importNode-null.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests that document.importNode(null) throws a TypeError");

shouldThrow("document.importNode(null)", "'TypeError: Type error'");
shouldThrow("document.importNode()", "'TypeError: Not enough arguments'");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
