LayoutTests/fast/dom/Document/importNode-null-expected.txt

Tests that document.importNode(null) throws a TypeError

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.importNode(null) threw exception TypeError: Type error.
PASS document.importNode() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
